refactor: 루트 레이아웃 위치 수정 및 스타일 수정
