+++ src/app/(without-header)/auth/layout.tsx
@@ -1,4 +1,3 @@
-import '../../globals.css'
 import type { ReactNode } from 'react'
 
 export default function AuthLayout({
@@ -7,20 +6,10 @@ export default function AuthLayout({
   children: ReactNode
 }) {
   return (
-    <html lang="ko">
-      <head>
-        <link
-          rel="stylesheet"
-          href="https://fastly.jsdelivr.net/gh/wanteddev/wanted-sans@v1.0.1/packages/wanted-sans/fonts/webfonts/variable/split/WantedSansVariable.min.css"
-        />
-      </head>
-      <body>
-        <div className="flex items-center justify-center font-sans min-h-screen bg-white md:bg-gray-100 text-gray-900">
-          <div className="py-8 px-6 border-gray-300 rounded-lg w-full max-w-md bg-white md:shadow-md">
-            {children}
-          </div>
-        </div>
-      </body>
-    </html>
+    <div className="fixed inset-0 flex items-center justify-center bg-white md:bg-gray-100">
+      <div className="py-8 px-6 border-gray-300 rounded-lg w-full max-w-md bg-white md:shadow-md">
+        {children}
+      </div>
+    </div>
   )
 }
